moved the utility funciton
