+++ src/mlstacks/constants.py
@@ -11,29 +11,6 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 """MLStacks constants."""
-import os
-
-
-def handle_bool_env_var(var: str, default: bool = False) -> bool:
-    """Converts normal env var to boolean.
-
-    Args:
-        var: The environment variable to convert.
-        default: The default value to return if the env var is not set.
-
-    Returns:
-        The converted value.
-    """
-    value = os.getenv(var)
-
-    if value in ["1", "y", "yes", "True", "true"]:
-        return True
-
-    if value in ["0", "n", "no", "False", "false"]:
-        return False
-
-    return default
-
 
 # enum for deployment types
 # deployment_type.KUBERNETES = "KUBERNETES"
